Isolate streaming Markdown from mobile
